Clarify method behavior
When `byteArraysMatch` is provided two non-zero offsets the call always
returns `false`. This is due to the `minSize` calculation being based
off of the length of both arrays. When this happens one of the offset
checks is guaranteed to always be greater than the length of itself.

So what is the intent?

It appears that the initial intent was possibly to take two arrays and
compare them. Since this method is an internal helper used in processing
a BLE data payload, as an array of bytes, it seems the offsets were
added as a way of avoiding creating the intermediate arrays. As the git
history is unclear this is simply an assumption; perhaps there were
performance concerns.

If we take this assumption as true then we can say the goal is to
compare the remainder of two arrays given a set of offsets. To achieve
this, keeping the existing method signature, we need to compare the
length of each array minus the respective offset. Then choose the
smaller of the two values:

```java
int len1 = array1.length - offset1;
int len2 = array2.length - offset2;
int minSize = len1 > len2 ? len2 : len1;
```

At this point I'd probably prefer to pull in `Math.min` for better
expressiveness and the ability to not declare the intermediate temp
length variables:

```java
int minSize = Math.min(array1.length - offset1, array2.length - offset2)
```

However, if we're going through all this trouble to be so flexible, why
not add a fifth parameter to explicitly state the length to compare?
This would be similar to how `System.arraycopy` works.

```java
byteArraysMatch(byte[] array1, int offset1, byte[] array2, int offset2, int length)
```

Not only does adding the length allow for a more narrowly defined
sub-array comparison, but it removes potential ambiguity with variable
size arrays. Without the length we simply check if the "smaller
subarray" exists within the larger starting at the given offset. The
"smaller subarray" could be either array. For example:

```java
byte[] a1 = { 1, 0, 1, 0 };
byte[] a2 = { 1, 0, 1 };
byteArraysMatch(a1, 0, a2, 0);    // true
byteArraysMatch(a1, 2, a2, 0);    // true
```

In the first comparison the bytes defined in `a2` are the "definitive
bytes". Whereas in the second comparison the sub-array of bytes from
`a1` are the "definitive bytes".

In the context of parsing BLE payloads we already know which array
should be the "definitive bytes". Leaving such ambiguity opens up the
potential for incorrectly matched payloads. Issues caused by this will
be costly to troubleshoot as they will likely only occur far downstream.
Thus the method signature will need to be changed to ensure correctness.

Since this is not a generic utility method we are free to implement the
more flexible method signature, or we could restrict it to just our use
case. As currently called, the value of `array1` is always our BLE
payload bytes. The value of `array2` is also always a predetermined set
of bytes expected to exist in their entirety in the BLE payload. This
usage has been static for some time and does not appear likely to
change. Based on the usage and context using a more restrictive method
signature produces more explicit code.

_NOTE: In case this needs to change in the future here is the more
flexible alternative implementation._

```java
private boolean byteArraysMatch(byte[] array1, int offset1, byte[] array2, int offset2, int length)
    if (array1.length - offset1 < length || array2.length - offset2 < length) {
        return false;
    }
    for (int i = 0; i <  length; i++) {
        if (array1[offset1 + i] != array2[offset2 + i]) {
            return false;
        }
    }
    return true;
}
```