Add changelog to project root
This adds a changelog file to the project root. This is a common file
many open source projects include. This provides a standard place for a
developer to look to get a quick summary of changes; say before an
upgrade.

Additionally, this type of file is GitHub agnostic. So if the project is
ever moved the change summary history is preserved. This also allows an
offline developer to see the summary while working on the project.

This changelog is formatted with markdown to provide a nice display in
the web UI. It also includes the pull request numbers for each of the
entries. This is great because it easily allows someone to look up the
associated code changes and any accompanying discussions.

Lastly, this includes the developer's name who did the work. Easily
accessible public credit is always a good thing.