Use elapsed time for scan stop time
All of the other time trackers in `CycledLeScanner` use the elapsed
time. I think it was simply an oversight when the change was made to
guard against the Android N (Android 7, API 24+) behavior of limiting
scan starts to only five every thirty seconds.

However, using the system time causes an issue where the scan will never
stop. This happens when the time is changed to the past or was set to
the future then return to normal or past time. Once `mLastScanStopTime`
is set if the system time ever goes backwards the calculation `now -
mLastScanStopTime` will always be negative. Thus it will always be less
than the `ANDROID_N_MIN_SCAN_CYCLE_MILLIS` prevent a natural stop.