Minor adjustments to `ScanData`
The `ScanData` class is `private` so it can only be used by the
`BeaconService` class. As there is no dependency relationship between
the classes this marks it as a `static` class. This saves on an internal
`BeaconService` instance reference in the `ScanData` object. Similarly,
since this is `private` the constructor is made package private. This
was suggested by Android Studio as functionally only the `BeaconService`
can instantiate an instance anyways.