Fix where scanner thread is quit.
This fixes what I believe was a simply copy-paste error regarding how
the scanner thread is quit. This uses the scanner thread handler,
instead of the main thread handler, to schedule the `quit` operation.
This way any queued scan jobs (such as a stop) complete before we quit
the thread. This is what the comment states, but wasn't what the code
implemented.

The `mHandler.removeCallbacksAndMessages` is removed from the background
job as `destroy` is already on the main thread. So we want to clear any
other jobs still in the queue to run later now as we are in the process
of destroying everything.