Fix possible NPE on older Android with power save
This eliminates the short circuit in the guard clause which would
prevent a `BeaconManager` from being set. The basic creation of a
`BeaconManager` instance doesn't require Android 18; running that
instance will and is turned into a no-op at that point. While here
nullability annotations were included and the `BeaconManager` instance
was made `final` (this is how the NPE was caught).