Add nullability annotations to `BeaconManager`
This chases down most instances of nullability within `BeaconManager`.
One notable omission was related to `distanceModelUpdateUrl`. As this is
`protected` it's accessible outside control of the class. So there's no
guarantee we can make either way about it's state. However, I didn't
want to add `@Nullable` to it as I'm not sure that's a valid condition.
Since I didn't know which way to go I left it alone.

There are a few places that once these annotations were added became
flagged by Android Studio. Some of these conditions were for methods
which are now annotated with `@NonNull` parameters but for which there
is a `null` check on it. In these particular situations while receiving
a `null` value is likely a programming error on the caller's part we
have acceptable behavior for how a `null` could be handled by
essentially being a no-op / falsey. These received a `noinspection`
"magic comment" telling Android Studio and lint to ignore them.

Finally, there were a few spots where Android Studio flagged possible
NPE errors regarding `sInstance` access. The reason being that in a
conditional expression `sInstance != null && sInstance.someMethod` is
not guaranteed to be atomic. So it's possible for `sInstance` to not be
`null` passing the first check but then be `null` when the method is
accessed. However, given how / what sets `sInstance` and where these
checks are the likelyhood of this happening is extremely low. Never the
less to ensure correctness this captures the value of `sInstance` at the
start of the call, and uses it for the checks, thus insuring validity of
the `null` check.