Remove excessive synchronization
The set of range and monitor notifiers is stored in a thread safe
`CopyOnWriteArraySet` instance. Adding an additional layer of
synchronization around already atomic operations is unnecessary. The
only case where we performed a non-atomic action is in the legacy
`getRangingNotifier` and `getMonitoringNotifier` implementations.

In these legacy cases, synchronization is necessary to ensure that the
empty set check isn't invalid by the following iterator access (which
would throw a `NoSuchElementException`). The iterator creation is
necessary because unlike lists sets don't allow elements to be requested
by index. But, if we drop the synchronization for all of the other use
cases, while leaving it here, doesn't provide us this guarantee.

However, we actually don't even have that guarantee now. This is because
the set is publicly exposed through the getters `getRangingNotifiers`
and `getMonitoringNotifiers`. Any client can request and manipulate the
sets external to the manager potentially causing a confusing
`NoSuchElementException` exception.

One of the reasons for using `CopyOnWriteArraySet` is that their
iterators are thread safe allowing unsynchronized callback iteration by
the beacon service. This drops the explicit synchronization and corrects
the legacy getter implementations by only using the iterator.