Wrap notifier sets before returning in API
This wraps the publicly returned notifier sets before exposing them
through the public API. This is necessary to maintain the integrity of
the internal sets. For the most part we don't need to do anything else
when a notifier is added or removed. However, exposing the set directly
gives up the ability to add more dependent behavior later (e.g. checks
such as `determineIfCalledFromSeparateScannerProcess` or even triggering
updates to the backing service state).

By using an "unmodifiable" set we accomplish four things:

- ensure changes can't be made directly to the notifiers outside the
  manager (see footnote)
- express the intent that this accessor is only for read access; instead
  of a simply copy which could be modified but won't actually change
  anything the manager uses
- produce a "fast" defensive "copy" of the set

  It is "fast" because the unmodifiable set doesn't actually create a
  new set; it simply wraps the existing set. This means no actual
  copying takes place either. Instead it "reads through" to the backing
  set. This means any changes made to the internal set will be visible
  to the unmodifiable set as well.
- provide a thread-safe view of the set

  The backing set is an instance of `CopyOnWriteArraySet` and guarantees
  all atomic read operations, and iterators, are thread-safe.

"New Instance" vs "Final Reader"
--------------------------------

Based on the "pass through" nature it's a fair question to ask:

> Why create a "new" set each time instead of maintaining a single
> internal "read-only" / "public" equivalent. For example:
>
> ```java
> final Set<Notifier> notifiers = new CopyOnWriteArraySet<>();
> final Set<Notifier> publicNotifiers =
>         Collections.unmodifiableSet(notifiers);
>
> public void addNotifier(Notifier notifier) {
>     if (notifier != null) {
>         notifiers.add(notifier);
>     }
> }
>
> public boolean removeNotifier(Notifier notifier) {
>     return notifiers.remove(notifier);
> }
>
> public Set<Notifier> getNotifiers() {
>     return publicNotifiers;
> }
> ```
>
> _Note: Class/object names changed for example brevity_

The short answer is to keep things simple and try to reduce cognitive
overhead/complexity. By having only a single set it's clear how things
work internally. The only time we need to be concerned with the
alternative "public" version is in the getter/any time it is exposed
publicly - which really should be minimized to the getter.

The performance differences between creating a "new instance" through
`Collections.unmodifiableSet` compared to returning a "constant" field
appear to negligible (verifying such benchmarks is left as an exercise
to the reader - sorry). It's possible by inlining the `unmodifiableSet`
call _could_ allow the compiler / JVM can further optimize it into
effectively a constant depending on the implementation in `Collections`;
which it may not be able to do when set as a field (but that's all
speculation on my part at the moment - but does make for an effective
argument :-D).

Footnote:

While this change _does_ prevent modification to the internal set
through the getter it doesn't guarantee outside modification. The
internal set is listed as `protected` access. So any subclass or object
in the same package could also modify the set. However, this does expose
the intent that the public API is meant to have a "read-only" level of
interaction.