Include basic copy set benchmarks
This adds some basic benchmarks for the performance of using
`Collections.unmodifiableSet` vs copying the actual set. This is setup
as an instrumented `androidTest` to allow for actual testing on Android
devices; compared to unit tests run on more powerful computers - which
also use a potentially different JVM.

Usage of these Android instrumentation tests does not affect the gradle
build. These types of tests are ignored by the normal gradle build
process.

This basic test setup was copied from a new Android Studio project. This
requires a change to the newer instrumentation runner suggested by
Google through Android Studio (we weren't using the old one). In order
to use this we need to include Espresso - which also then requires us to
change the min SDK. To keep the old SDK version, but allow these
instrumented tests to run with Espresso we include test only options in
the build config.