Include running average in Parcel data
This fixes a minor issue where the running average is not set after
restoring a beacon from a parcel. Looking at the history it seems that
the running average was an internal detail until #479 where support was
added for running the service in a separate process.

Part of the reason this needs to be included is that on Android 8.0 the
running average _is_ restored and available for access. The reason the
behavior is different across Android versions is in the way a service
intent versus local notification handles the data `Bundle` (i.e. which
[`Callback`](https://github.com/AltBeacon/android-beacon-library/blob/2.12.3/src/main/java/org/altbeacon/beacon/service/Callback.java#L58-L86)
branch runs). Explicitly setting the `BeaconManager` to use scheduled
scan jobs, forcing the first conditional code path using local
notifications, allows the running average to be available on older
versions.