Delegate `AltBeacon` constructors to parent
It's unclear from the history why only some of the fields are part of
the `AltBeacon` constructor. When initially implemented the constructor
omitted copying `mRunningAverageRssi` and `mManufacturer`. Since then
several additional data fields were added which are also not in the
`AltBeacon` constructor:

  - `mExtraDataFields`
  - `mServiceUuid`
  - `mBluetoothName`
  - `mParserIdentifier`
  - `mMultiFrameBeacon`

The only real change in this subclass seems to be the additional custom
getter `getMfgReserved`. Based on the fact that the parceling logic and
the builder all delegate to `Beacon` it seems reasonable that the
general intended logic was to wrap the parent class.

This fixes the two constructors so that all setup is based on the
`Beacon` implementation. This ensures that any changes in `Beacon` are
reflected by `AltBeacon` as well.