Add omitted fields in copy beacon constructor
It's unclear why `mManufacturer` wasn't included with the copy
constructor originally. However, as it's an internal field which is also
included in the serialization and parceling it probably should be copied
here too.

The `mMultiFrameBeacon` field was added later to provide better support
for beacons with extra data packets (#387). It's likely this was just
overlooked as most of the "copying" of beacons occurred through
parceling at the time.