Use annotations and final fields to express intent
Based on the code the implicit expectation is that the internal map will
never be `null`. Additionally, the provided `Beacon` to track should
also never be `null`. Adding the annotations helps let the Android
Studio code inspector and the lint tools know about this. This can help
prevent accidental NPE issues if someone can produce a `null` beacon
which then gets passed here.

While the map may change the variable is never re-assigned. Looking at
the code which uses this class when it needs to change the flag the
clients simply creates a new instance. Marking the fields `final` makes
this behavior intent clear.