Fix up extra data beacon test cases
This adjusts the test cases to ensure all of the branches get tested.
The `gattBeaconExtraDataGetUpdated` test is updated to show that each
GATT data beacon causes the tracked beacon to be updated. This was
implicitly assumed previously but this makes it explicit by re-checked
after each tracking call.

This removes the `gattBeaconExtraDataAreNotOverwritten` test. This test
is confusing. It is worded to indicate a change shouldn't happen even
though a change does occur to `beacon` / `trackedBeacon`. Since a change
occurs to `beacon` / `trackedBeacon` which is used in the assertion we
cannot say this proves `extraDataBeacon` wasn't modified.

This adds a new test `gattBeaconFieldsAreNotUpdated` to demonstrate that
without the GATT data beacons no changes to tracked beacons occur.
Conversely, this updates `gattBeaconFieldsGetUpdated` to prove that the
non-data beacons are updated based on the stored tracked beacons. It
uses a new "repeat" beacon to show that the extra data was modified, but
the main data and RSSI were left as reported by that beacon. The repeat
beacon is required because the original `beacon` is mutated and thus
would incorrectly show changes to the RSSI and data.