Minor refactor to improve separate update branches
This extracts the GATT data beacon path into a guard clause. This helps
to highlight the fact that these beacons:

  - are not modified
  - update any existing tracked beacons
  - return `null`

This also has the benefit or removing the conditional recheck within the
loop. This also makes it clear that we update the extra data based on
any one of the tracked beacons (as they all have the same data); instead
of resetting within the loop.

The tracking hash update logic is now inline within the main
`trackGattBeacon` method. It was only called from this method and only
for non-GATT data beacons. Also, since the GATT data logic is split out
we still need to set the extra data. Accessing the tracked beacons once
here keeps all of this logic together making it easier to see what is
happening on an non-data beacon update.