Help protected against NPE for extra data tracking
This is a follow up to #616, #617, and #626. It uses annotations to
indicate that we never expect this field to be `null`. It also sets a
default tracker on initialization. This ensures there is always at
tracker available; even if it's only temporary. The work in the other
PRs makes sure the tracker is re-initialized at the appropriate times with
the correct settings.