Samsung devices stop scanning when screen is off
We have been encountering an issue where Android 9 Samsung devices are scanning with the screen is on, with wildcard scanning, then once the screen is off it doesn't switch over to non-wildcard scanning immediately, causing gaps in scanning to take place. 

Because the current code checks if it is interactive, it doesn't fall into non-wildcard scanning until the screen is off. By immediately going into non-wildcard scanning, we prevent gaps during the transition period. 

We were also able to create a work around the stops ranging/starts ranging for regions when the screen turns off